+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/util/SpotbugsUtils.groovy
@@ -7,7 +7,6 @@ import groovy.xml.XmlUtil
 import org.gradle.api.Project
 import org.gradle.api.tasks.SourceSet
 import org.gradle.api.tasks.compile.JavaCompile
-import org.gradle.util.GradleVersion
 import org.slf4j.Logger
 import ru.vyarus.gradle.plugin.quality.QualityExtension
 
@@ -84,9 +83,7 @@ class SpotbugsUtils {
         }
         // apt is a special dir, not mentioned in sources!
         JavaCompile javaCompile = task.project.tasks.findByName(set.compileJavaTaskName) as JavaCompile
-        File aptGenerated = GradleVersion.current() < GradleVersion.version('7.0')
-                ? javaCompile.options.annotationProcessorGeneratedSourcesDirectory
-                : javaCompile.options.generatedSourceOutputDirectory.get().asFile
+        File aptGenerated = javaCompile.options.generatedSourceOutputDirectory.get().asFile
 
         Set<File> ignored = FileUtils.resolveIgnoredFiles(task.sourceDirs.asFileTree, extension.exclude)
         // exclude all apt-generated files