+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/ExcludeSourcesKitTest.groovy
@@ -114,11 +114,13 @@ class ExcludeSourcesKitTest extends AbstractKitTest {
         build("""
             plugins {
                 id 'java'
-                id 'ru.vyarus.animalsniffer' version '1.5.0'
+                id 'ru.vyarus.animalsniffer' version '1.7.1'
                 id 'ru.vyarus.quality'
             }
             
-            sourceCompatibility = 1.8
+            java {
+                sourceCompatibility = 1.8
+            }
 
             sourceSets.main {
                 java {