+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/MultiModuleUseKitTest.groovy
@@ -35,19 +35,16 @@ class MultiModuleUseKitTest extends AbstractKitTest {
         file('settings.gradle') << "include 'mod1', 'mod2'"
 
         fileFromClasspath('mod1/src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
-        fileFromClasspath('mod2/src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
+        fileFromClasspath('mod2/src/main/java/sample/Sample2.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample2.java')
 
         when: "run check for both modules"
         BuildResult result = run('check')
 
         then: "violations detected in module only"
-        unifyString(result.output).replaceAll("Total time: .*", "")
+        def out =  unifyString(result.output).replaceAll("Total time: .*", "")
                 // only on jdk 8
                 .replaceAll("WARNING: checkstyle-backport-jre8 .*\n", "")
-                .trim().startsWith("""> Task :mod1:compileJava
-> Task :mod1:processResources NO-SOURCE
-> Task :mod1:classes
-
+                out.contains("""
 > Task :mod1:checkstyleMain
 Checkstyle rule violations were found. See the report at: file:///tmp/junit6300057182805361069/mod1/build/reports/checkstyle/main.html
 Checkstyle files with violations: 1
@@ -64,16 +61,9 @@ Checkstyle violations by severity: [error:2]
   Missing a Javadoc comment.
   http://checkstyle.sourceforge.io/config_javadoc.html#MissingJavadocType
 
+""".replaceAll("tmp/junit6300057182805361069", ReportUtils.noRootFilePath(testProjectDir)))
 
-> Task :mod1:compileTestJava NO-SOURCE
-> Task :mod1:processTestResources NO-SOURCE
-> Task :mod1:testClasses UP-TO-DATE
-> Task :mod1:test NO-SOURCE
-> Task :mod1:check
-> Task :mod2:compileJava
-> Task :mod2:processResources NO-SOURCE
-> Task :mod2:classes
-
+        out.contains("""
 > Task :mod2:checkstyleMain
 Checkstyle rule violations were found. See the report at: file:///tmp/junit6300057182805361069/mod2/build/reports/checkstyle/main.html
 Checkstyle files with violations: 1
@@ -82,21 +72,14 @@ Checkstyle violations by severity: [error:2]
 
 2 Checkstyle rule violations were found in 1 files
 
-[Misc | NewlineAtEndOfFile] sample.(Sample.java:1)
+[Misc | NewlineAtEndOfFile] sample.(Sample2.java:1)
   File does not end with a newline.
   http://checkstyle.sourceforge.io/config_misc.html#NewlineAtEndOfFile
 
-[Javadoc | MissingJavadocType] sample.(Sample.java:6)
+[Javadoc | MissingJavadocType] sample.(Sample2.java:6)
   Missing a Javadoc comment.
   http://checkstyle.sourceforge.io/config_javadoc.html#MissingJavadocType
 
-
-> Task :mod2:compileTestJava NO-SOURCE
-> Task :mod2:processTestResources NO-SOURCE
-> Task :mod2:testClasses UP-TO-DATE
-> Task :mod2:test NO-SOURCE
-> Task :mod2:check
-
-BUILD SUCCESSFUL""".replaceAll("tmp/junit6300057182805361069", ReportUtils.noRootFilePath(testProjectDir)))
+""".replaceAll("tmp/junit6300057182805361069", ReportUtils.noRootFilePath(testProjectDir)))
     }
 }