+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/tools/animalsniffer/AnimalSnifferIntegrationKitTest.groovy
@@ -15,11 +15,13 @@ class AnimalSnifferIntegrationKitTest extends AbstractKitTest {
         build("""
             plugins {
                 id 'java'
-                id 'ru.vyarus.animalsniffer' version '1.5.0'
+                id 'ru.vyarus.animalsniffer' version '1.7.1'
                 id 'ru.vyarus.quality'
             }
             
-            sourceCompatibility = 1.8
+            java {
+                sourceCompatibility = 1.8
+            }
 
             quality {
                 strict false
@@ -52,7 +54,7 @@ class AnimalSnifferIntegrationKitTest extends AbstractKitTest {
         build("""
             plugins {
                 id 'java'
-                id 'ru.vyarus.animalsniffer' version '1.5.0'
+                id 'ru.vyarus.animalsniffer' version '1.7.1'
                 id 'ru.vyarus.quality'
             }